requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
flask-session~=0.3.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msal==1.8.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
urllib3==1.26.2
werkzeug<1
